requirements.txt
flask
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
paho-mqtt
requests
gunicorn